+++ app/server/controllers/active-cases.ts
@@ -3,7 +3,7 @@ import * as Paths from '../paths';
 import * as AppInsights from '../app-insights';
 import { Logger } from '@hmcts/nodejs-logging';
 import { getCasesByName } from '../utils/fieldValidation';
-import { CaseDetails } from '../data/models';
+import { CaseDetails } from 'app/server/models/express-session';
 import { Dependencies } from '../routes';
 import { isCaseActive } from './cases';
