+++ app/server/middleware/setLanguage.ts
@@ -4,14 +4,12 @@ import {
   NextFunction,
   Router as expressRouter,
 } from 'express';
-import * as config from 'config';
+import config from 'config';
 import { Logger } from '@hmcts/nodejs-logging';
 import { LoggerInstance } from 'winston';
 import { resolveQuery } from '../utils/parseUtils';
 
 import i18next from 'i18next';
-
-const i18n = require('i18next');
 
 const languages: Array<string> = config.get('languages');
 
@@ -32,7 +30,7 @@ export async function setLanguage(
   } else {
     await i18next.changeLanguage(sessionLanguage);
   }
-  logger.info(`Language is set to ${i18n.language}`);
+  logger.info(`Language is set to ${i18next.language}`);
   next();
 }
