+++ test/unit/controllers/task-list.test.ts
@@ -11,7 +11,7 @@ import * as Paths from 'app/server/paths';
 import { Feature } from 'app/server/utils/featureEnabled';
 import { expect, sinon } from '../../chai-sinon';
 import { INTERNAL_SERVER_ERROR } from 'http-status-codes';
-import { CaseDetails } from 'app/server/data/models';
+import { CaseDetails } from 'app/server/models/express-session';
 import { Dependencies } from 'app/server/routes';
 
 describe('controllers/task-list', function () {
